+++ client/src/pages/Component/Visualizer/Entities/BaseEntity/baseEntity.tsx
