+++ client/src/pages/Component/Visualizer/Entities/BaseEntity/edgeEntity.tsx
