+++ client/src/pages/Component/Visualizer/Entities/BaseEntity/nodeEntity.tsx
