+++ client/src/pages/Component/Visualizer/Entities/BaseEntity/pointerEntity.tsx
