+++ client/src/pages/Component/Visualizer/Entities/CoreEntity/attachableEntity.ts
@@ -1,8 +1,8 @@
-import { assertUnreachable } from '../../Component/Visualizer/Util/util';
-import { EntityType } from '../Entity/baseEntity';
-import { NodeEntity } from '../Entity/nodeEntity';
-import { EntityConcrete } from '../frontendType';
-import { Circle, ShapeType } from '../Geometry/geometry';
+import { assertUnreachable } from '../../Util/util';
+import { EntityType } from '../BaseEntity/baseEntity';
+import { NodeEntity } from '../BaseEntity/nodeEntity';
+import { EntityConcrete } from '../../../../Types/frontendType';
+import { Circle, ShapeType } from '../../../../Types/geometryType';
 
 /**
  * Attachable Entity refers to the type of entity that has a shape without infers from other entity.
